+++ BxInputController/Sources/Common/View/BxInputBaseRowBinder.swift
@@ -56,6 +56,14 @@ open class BxInputBaseRowBinder<Row: BxInputRow, Cell : UITableViewCell> : NSObj
     
     public var checkers: [BxInputRowChecker] = []
     
+    public var needChangeDisadledCell: Bool {
+        return owner?.settings.isNormalShowingDisadledCell == false
+    }
+    
+    public var alphaForDisabledView: CGFloat {
+        return owner?.settings.alphaForDisabledView ?? 0.5
+    }
+    
     /// new value should map with data model
     public init(row: Row)
     {