+++ CHANGELOG.md
@@ -6,6 +6,14 @@ ByteriX, 2017. All right reserved.
 
 # Versions
 
+## 2.4.0 (20.07.2017)
+##### Features
+* new methods setEnableSection, setEnable for batching process of dissable/enable rows
+* add to setEnableRow param for animating
+* new methods reloadSections for reload array of sections
+* new params in Settings for managment dissable/enable rows: isNormalShowingDisadledCell, changeViewEnableHandler, alphaForDisabledView
+* edit example project for showing the changes regarging dissable/enable rows
+
 ## 2.3.11 (22.06.2017)
 ##### Improvements
 * fixed placeholder color from formatted text